fix: Final BootstrapUtils split into HandlebarsUtils
