+++ src/service/index.ts
@@ -6,7 +6,6 @@ export * from './AnnounceService';
 export * from './AsyncUtils';
 export * from './BootstrapAccountResolver';
 export * from './BootstrapService';
-export * from './BootstrapUtils';
 export * from './CertificateService';
 export * from './CommandUtils';
 export * from './ComposeService';
@@ -16,6 +15,8 @@ export * from './ConfigurationUtils';
 export * from './Constants';
 export * from './CryptoUtils';
 export * from './FileSystemService';
+export * from './HandlebarsUtils';
+export * from './KnownError';
 export * from './LinkService';
 export * from './MigrationService';
 export * from './ModifyMultisigService';
