+++ test/service/Utils.test.ts
@@ -34,4 +34,68 @@ describe('Utils', () => {
             'Running image using Exec: symbolplatform/symbol-server:tools-gcc-0.10.0.5 /usr/catapult/bin/catapult.tools.votingkey --secret=HIDDEN_KEY --startEpoch=1 --endEpoch=26280 --output=/votingKeys/private_key_tree1.dat\n',
         );
     });
+
+    it('should remove null values', () => {
+        const compose = {
+            version: '2.4',
+            services: {
+                db: {
+                    user: '',
+                    environment: {
+                        MONGO_INITDB_DATABASE: 'null',
+                    },
+                    container_name: 'db',
+                    image: 'mongo:4.4.3-bionic',
+                    command: 'mongod --dbpath=/dbdata --bind_ip=db',
+                    stop_signal: 'SIGINT',
+                    working_dir: '/docker-entrypoint-initdb.d',
+                    ports: [],
+                    volumes: ['./mongo:/docker-entrypoint-initdb.d:ro', '../databases/db:/dbdata:rw'],
+                    mem_limit: null,
+                },
+                networks: {
+                    default: {
+                        ipam: {
+                            config: [
+                                {
+                                    subnet: '172.20.0.0/24',
+                                },
+                            ],
+                        },
+                    },
+                },
+            },
+        };
+
+        const composePruned = {
+            version: '2.4',
+            services: {
+                db: {
+                    user: '',
+                    environment: {
+                        MONGO_INITDB_DATABASE: 'null',
+                    },
+                    container_name: 'db',
+                    image: 'mongo:4.4.3-bionic',
+                    command: 'mongod --dbpath=/dbdata --bind_ip=db',
+                    stop_signal: 'SIGINT',
+                    working_dir: '/docker-entrypoint-initdb.d',
+                    volumes: ['./mongo:/docker-entrypoint-initdb.d:ro', '../databases/db:/dbdata:rw'],
+                },
+                networks: {
+                    default: {
+                        ipam: {
+                            config: [
+                                {
+                                    subnet: '172.20.0.0/24',
+                                },
+                            ],
+                        },
+                    },
+                },
+            },
+        };
+
+        expect(Utils.pruneEmpty(compose)).to.deep.eq(composePruned);
+    });
 });
